add streaming asr demo, test=doc
