+++ demos/streaming_asr_server/run.sh
@@ -1,0 +1,2 @@
+# start the streaming asr service
+paddlespeech_server start --config_file ./conf/ws_conformer_application.yaml
