+++ demos/streaming_asr_server/web/paddle_web_demo.png
