+++ demos/streaming_asr_server/web/static/js/SoundRecognizer.js
