+++ demos/streaming_asr_server/web/static/js/recorder/extensions/frequency.histogram.view.js
