+++ demos/streaming_asr_server/web/static/js/recorder/recorder-core.js
